requirements.txt
# To install the required packages, run the following command:
# pip install -r requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
aiohttp>=3.9.1